backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq==0.11.0
python-dotenv==1.0.1
sse-starlette==2.1.3
